src/app/questions/mortgage-intent/buying/core/confirm-location-details/confirm-location-details.guard.spec.ts
import {
  TestBed,
  inject
} from '@angular/core/testing';

import { CanActivateQuestionsMortgageIntentConfirmLocation } from './confirm-location-details.guard';
import { MortgageIntentBuyingRoutingService } from '../buying-routing.service';

describe('CanActivateQuestionsMortgageIntentConfirmLocationDetails', () => {
  let guard;
  let service;
  beforeEach(() => {
    TestBed.configureTestingModule({
      providers: [
        CanActivateQuestionsMortgageIntentConfirmLocation,
        {
          provide: MortgageIntentBuyingRoutingService,
          useValue: jasmine.createSpyObj('routingService', ['validConfirmLocation'])
        }
      ]
    });
  });

  beforeEach(inject([MortgageIntentBuyingRoutingService], (_service_) => {
    service = _service_;
  }));

  beforeEach(inject([CanActivateQuestionsMortgageIntentConfirmLocation], (_guard_) => {
    guard = _guard_;
  }));

  it('should create', () => {
    expect(guard).toBeTruthy();
  });

  it('should have the canActivate method', () => {
    expect(guard.canActivate).toEqual(jasmine.any(Function));
  });

  describe('canActivate', () => {
    it('should return routingService.validConfirmLocation', () => {
      service.validConfirmLocation.and.returnValue(Promise.resolve(true));
      guard.canActivate().then(bool => {
        expect(bool).toEqual(true);
        expect(service.validConfirmLocation).toHaveBeenCalledWith();
        expect(service.validConfirmLocation).toHaveBeenCalledTimes(1);
      });

      service.validConfirmLocation.calls.reset();
      service.validConfirmLocation.and.returnValue(Promise.resolve(false));
      guard.canActivate().then(bool => {
        expect(bool).toEqual(false);
        expect(service.validConfirmLocation).toHaveBeenCalledWith();
        expect(service.validConfirmLocation).toHaveBeenCalledTimes(1);
      });
    });
  });
});
